Fix various issues

+++ include/alpaka/mem/buf/BufGenericSycl.hpp
@@ -221,7 +221,7 @@ namespace alpaka
                     auto const width = extent::getWidth(ext);
                     auto const widthBytes = width * static_cast<TIdx>(sizeof(TElem));
 
-                    memPtr = malloc_device<TElem>(width, dev.m_device, dev.m_context);
+                    memPtr = malloc_device<TElem>(static_cast<std::size_t>(width), dev.m_device, dev.m_context);
                     pitchBytes = static_cast<TIdx>(widthBytes);
 
 #if ALPAKA_DEBUG >= ALPAKA_DEBUG_FULL
@@ -238,7 +238,7 @@ namespace alpaka
                     auto const widthBytes = width * static_cast<TIdx>(sizeof(TElem));
                     auto const height = extent::getHeight(ext);
 
-                    memPtr = malloc_device<TElem>(width * height, dev.m_device, dev.m_context);
+                    memPtr = malloc_device<TElem>(static_cast<std::size_t>(width * height), dev.m_device, dev.m_context);
                     pitchBytes = static_cast<TIdx>(widthBytes);
 
 #if ALPAKA_DEBUG >= ALPAKA_DEBUG_FULL
@@ -258,7 +258,7 @@ namespace alpaka
                     auto const height = extent::getHeight(ext);
                     auto const depth = extent::getDepth(ext);
 
-                    memPtr = malloc_device<TElem>(width * height * depth, dev.m_device, dev.m_context);
+                    memPtr = malloc_device<TElem>(static_cast<std::size_t>(width * height * depth), dev.m_device, dev.m_context);
                     pitchBytes = static_cast<TIdx>(widthBytes);
 
 #if ALPAKA_DEBUG >= ALPAKA_DEBUG_FULL
